Fix Jedi REPL completion in IPython
For whatever reason, IPython resets the readline completer to its own
completer after each prompt, which was preventing the Jedi completer from
working.  This checks if the Python is IPython and disables the IPython
completer.

This also adds an experimental option (off by default) to combine the old
completer and the Jedi completer, which (if it works) should allow to fallback
to the IPython completer when the Jedi one fails.  So far it doesn't seem to
work, which is why I disabled it.