Dynamic check for VIRTUAL_ENV python version
Previously, an assumption was made that the virtualenv was created using
a python version that matched sys.version_info[:2]. If this was not the
case, an invalid path is added to sys.path and we don't get any
completion.

This commit checks the 'pythonX.X' file within $VIRTUAL_ENV/lib and sets
the path accordingly. The test_get_sys_path() test has been updated to
properly test this by mocking os.listdir().