docstrings: fix "Sphinx param with type" pattern
Previously, the pattern only matched if the parameter description
followed on the same line, like so:

    :param type foo: A param named foo.

However, it's also valid for the parameter description to be wrapped
onto the next line, like so:

    :param type foo:
        A param named foo.

This change updates the pattern to match the second example as well, and
adds a test to verify this behaviour.

Fixes #806.