Some performance enhancement.
- Make sure the `CachedFastParser` are cached on disk
- Change the pickle serialisation on dist is optimized **on write**,
this does not change the pickling version, it just make it more
efficient (unless things have already been written)
- Do not trust the `SequenceMatcher` to be efficient if the two files
are the same, just take a shortcut and return previous completer.
- Do some argument caching in the core of the parser loop.

With this the 2 remaining method that could be made faster and are
taking most of the time are:
 - the `splitlines` one, but hard to be made much faster
 - the `pickle.load` one, if we come up with our own serialisation.

With this, on my machine I have in some case a ~15 fold increase in
speed. Leading to things like `matplotlib.<tab>` becomming reactive
enough to use.