Make better use of Parser Caching.
As far as I can tell, this should make a better use of existing parser
for modules. This is the core of what get the inference of types for
some completion on my machine from ~20s to ~2s. We are still caching on
disk  and it should be possible to cache more agressively on other
locations.