Increase Python's recursion limit
Currently there is still the possiblity that Jedi fails with a recursion error,
because the stack is too small. (see #861) By increasing it we improve the
situation.

Probably we should just be switching away from this extreme amount of recursion
and move to queueing which would also allow to use other algorithms such as
breadth-first-search.