Remove the old definition tests
The reason for this is that they haven't been used in years and don't really
make sense, because the way we now resolve parentheses is by executing the
result.

IMO this was a good patch at the time, but doesn't make sense anymore. Let me
know if you disagree ~dave.